src/findings-writer/requirements.txt
Flask==2.0.1
google-cloud-bigquery==2.23.3
google-cloud-storage==1.41.1
gunicorn==20.1.0